Fix implicit dependency in 'System' library
One or more functions in the 'System' library make use of functions in
the 'Log' library. To prevent errors from happen, the user would have
to know about this implicit and include the 'Log' library too.
In order to simplify the use, implicit dependencies like this will now
be sourced in the libraries directly. That way no knowledge of implicit
dependencies is required.