Improve documentation, samples, and templates
Revamp/replace/add samples and templates. We should have two templates,
one for a library function and one for an entire script. We should also
have two examples, one for a library function and one for an entire
script.

Updates NOTES and README as needed. NOTES in particular needs a lot
of attention.

Update function documentation as well.